Move the yaml conversions over to ecs
