+++ Sprocket/CMakeLists.txt
@@ -45,6 +45,7 @@ add_library(Sprocket STATIC
             Scene/scene.cpp
             Scene/loader.cpp
             Scene/lua_ecs.cpp
+            Scene/yaml_ecs.cpp
             
             Graphics/Animation.cpp
             Graphics/buffer_element_types.cpp
