+++ Sprocket/Scene/loader.dm.cpp
@@ -1,7 +1,7 @@
 #include "loader.h"
 #include "meta.h"
 #include "ecs.h"
-#include "lua_ecs.h"
+#include "yaml_ecs.h"
 
 #include <Sprocket/Core/log.h>
 #include <Sprocket/Utility/Yaml.h>
